+++ src/main/java/de/maxhenkel/voicechat/voice/client/GroupChatManager.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class GroupChatManager {
 
     private static final ResourceLocation TALK_OUTLINE = new ResourceLocation(Voicechat.MODID, "textures/gui/talk_outline.png");
-    private static final ResourceLocation SPEAKER_OFF_ICON = new ResourceLocation(Voicechat.MODID, "textures/gui/speaker_off.png");
+    private static final ResourceLocation SPEAKER_OFF_ICON = new ResourceLocation(Voicechat.MODID, "textures/gui/speaker_group_hud_small_off.png");
 
     public static void renderIcons(PoseStack matrixStack) {
         Client client = VoicechatClient.CLIENT.getClient();
@@ -55,8 +55,8 @@ public class GroupChatManager {
 
             if (state.isDisabled()) {
                 matrixStack.pushPose();
-                matrixStack.translate(10, 5, 0);
-                matrixStack.scale(0.25F, 0.25F, 1F);
+                matrixStack.translate(10, 1, 0);
+                matrixStack.scale(0.5F, 0.5F, 1F);
                 RenderSystem.setShader(GameRenderer::getPositionTexShader);
                 RenderSystem.setShaderColor(1F, 1F, 1F, 1F);
                 RenderSystem.setShaderTexture(0, SPEAKER_OFF_ICON);