+++ changelog.md
@@ -1,2 +1,2 @@
-- Fixed crash when using special characters as group names
-- Fixed group invitations not working with spaces in the group name
+- Fixed optional icon resource pack format
+- Fixed group hud icons